Extract useRankedScan: one copy of the scan whose subtleties are load-bearing
FilesHome had a verbatim copy of useWalkSearch's scan block — score/sort/
setTimer/clearTimer/onPublish, the query-tagged `scanned` state, and the
`scanned.q === q` gate. Every one of those is a fix for a specific bug (the
tag must be committed WITH the data or the old query's rows get labelled with
the new one; `done` must survive into state or an intermediate slice reads as
a settled "no matches"), and scan-job.test.ts guards them by reading the
source — which it could only do for one of the two copies.

The copy had already diverged: it passed `onProgress: () => {}`, so a scan
cancelled mid-flight on the home page restarted from zero instead of resuming
where it left off. Sharing the hook fixes that by construction.

The only real difference between the callers was the debounce — the in-folder
walk arrives in stream flushes, the home corpus arrives once — so that is a
parameter. The source guard now points at the single consumer.
